requirements.txt
streamlit
plotly.express
pandas
sqlite3
